partition code between resnet_run_loop and resnet_model
