+++ official/resnet/cifar10_main.py
@@ -23,7 +23,8 @@ import sys
 
 import tensorflow as tf
 
-from official.resnet import resnet
+from official.resnet import resnet_model
+from official.resnet import resnet_run_loop
 
 _HEIGHT = 32
 _WIDTH = 32
@@ -126,22 +127,22 @@ def input_fn(is_training, data_dir, batch_size, num_epochs=1,
 
   num_images = is_training and _NUM_IMAGES['train'] or _NUM_IMAGES['validation']
 
-  return resnet.process_record_dataset(dataset, is_training, batch_size,
-      _NUM_IMAGES['train'], parse_record, num_epochs, num_parallel_calls,
-      examples_per_epoch=num_images, multi_gpu=multi_gpu)
+  return resnet_run_loop.process_record_dataset(dataset, is_training, batch_size,
+                                                _NUM_IMAGES['train'], parse_record, num_epochs, num_parallel_calls,
+                                                examples_per_epoch=num_images, multi_gpu=multi_gpu)
 
 
 def get_synth_input_fn():
-  return resnet.get_synth_input_fn(_HEIGHT, _WIDTH, _NUM_CHANNELS, _NUM_CLASSES)
+  return resnet_run_loop.get_synth_input_fn(_HEIGHT, _WIDTH, _NUM_CHANNELS, _NUM_CLASSES)
 
 
 ###############################################################################
 # Running the model
 ###############################################################################
-class Cifar10Model(resnet.Model):
+class Cifar10Model(resnet_model.Model):
 
   def __init__(self, resnet_size, data_format=None, num_classes=_NUM_CLASSES,
-      version=resnet.DEFAULT_VERSION):
+      version=resnet_model.DEFAULT_VERSION):
     """These are the parameters that work for CIFAR-10 data.
 
     Args:
@@ -180,7 +181,7 @@ def cifar10_model_fn(features, labels, mode, params):
   """Model function for CIFAR-10."""
   features = tf.reshape(features, [-1, _HEIGHT, _WIDTH, _NUM_CHANNELS])
 
-  learning_rate_fn = resnet.learning_rate_with_decay(
+  learning_rate_fn = resnet_run_loop.learning_rate_with_decay(
       batch_size=params['batch_size'], batch_denom=128,
       num_images=_NUM_IMAGES['train'], boundary_epochs=[100, 150, 200],
       decay_rates=[1, 0.1, 0.01, 0.001])
@@ -197,26 +198,26 @@ def cifar10_model_fn(features, labels, mode, params):
   def loss_filter_fn(name):
     return True
 
-  return resnet.resnet_model_fn(features, labels, mode, Cifar10Model,
-                                resnet_size=params['resnet_size'],
-                                weight_decay=weight_decay,
-                                learning_rate_fn=learning_rate_fn,
-                                momentum=0.9,
-                                data_format=params['data_format'],
-                                version=params['version'],
-                                loss_filter_fn=loss_filter_fn,
-                                multi_gpu=params['multi_gpu'])
+  return resnet_run_loop.resnet_model_fn(features, labels, mode, Cifar10Model,
+                                         resnet_size=params['resnet_size'],
+                                         weight_decay=weight_decay,
+                                         learning_rate_fn=learning_rate_fn,
+                                         momentum=0.9,
+                                         data_format=params['data_format'],
+                                         version=params['version'],
+                                         loss_filter_fn=loss_filter_fn,
+                                         multi_gpu=params['multi_gpu'])
 
 
 def main(unused_argv):
   input_function = FLAGS.use_synthetic_data and get_synth_input_fn() or input_fn
-  resnet.resnet_main(FLAGS, cifar10_model_fn, input_function)
+  resnet_run_loop.resnet_main(FLAGS, cifar10_model_fn, input_function)
 
 
 if __name__ == '__main__':
   tf.logging.set_verbosity(tf.logging.INFO)
 
-  parser = resnet.ResnetArgParser()
+  parser = resnet_run_loop.ResnetArgParser()
   # Set defaults that are reasonable for this model.
   parser.set_defaults(data_dir='/tmp/cifar10_data',
                       model_dir='/tmp/cifar10_model',
